refactor(claude): move semantic daemon behind adapter boundary
